requirement.txt
streamlit 
torch 
torchvision
pillow
